:recycle: Separate useScrollHeader
